lkl: make rumprun nolibc working again

+++ buildrump.sh
@@ -432,7 +432,7 @@ maketoolwrapper ()
 	else
 		lcx=$(echo ${tool} | tr '[A-Z]' '[a-z]')
 	fi
-	tname=${BRTOOLDIR}/bin/${MACHINE_GNU_ARCH}--netbsd${TOOLABI}-${lcx}
+	tname=${BRTOOLDIR}/bin/${MACHINE_GNU_ARCH}--${RUMPKERNEL}${TOOLABI}-${lcx}
 
 	printoneconfig 'Tool' "${tool}" "${fptool}"
 
@@ -533,7 +533,7 @@ maketools ()
 	else
 		cppname=cpp
 	fi
-	tname=${BRTOOLDIR}/bin/${MACHINE_GNU_ARCH}--netbsd${TOOLABI}-${cppname}
+	tname=${BRTOOLDIR}/bin/${MACHINE_GNU_ARCH}--${RUMPKERNEL}${TOOLABI}-${cppname}
 	printf '#!/bin/sh\n\nexec %s -E -x c "${@}"\n' ${CC} > ${tname}
 	chmod 755 ${tname}
 
@@ -548,7 +548,8 @@ maketools ()
 	${HOST_CC} -o ${BRTOOLDIR}/bin/brrealpath \
 	    ${BRDIR}/brlib/utils/realpath.c || die failed to build brrealpath
 
-	cat >> "${MKCONF}" << EOF
+	if [ ${RUMPKERNEL} = "netbsd" ] ; then
+		cat >> "${MKCONF}" << EOF
 BUILDRUMP_IMACROS=${BRIMACROS}
 .if \${BUILDRUMP_SYSROOT:Uno} == "yes"
 BUILDRUMP_CPPFLAGS=--sysroot=\${BUILDRUMP_STAGE}
@@ -569,13 +570,14 @@ RUMP_NPF_TESTING?=no
 RUMPRUN=yes
 EOF
 
-	if ! ${KERNONLY}; then
-		# queue.h is not available on all systems, but we need it for
-		# the hypervisor build.  So, we make it available in tooldir.
-		mkdir -p ${BRTOOLDIR}/compat/include/sys \
-		    || die create ${BRTOOLDIR}/compat/include/sys
-		cp -p ${SRCDIR}/sys/sys/queue.h ${BRTOOLDIR}/compat/include/sys
-		echo "CPPFLAGS+=-I${BRTOOLDIR}/compat/include" >> "${MKCONF}"
+		if ! ${KERNONLY}; then
+		    # queue.h is not available on all systems, but we need it for
+		    # the hypervisor build.  So, we make it available in tooldir.
+		    mkdir -p ${BRTOOLDIR}/compat/include/sys \
+			|| die create ${BRTOOLDIR}/compat/include/sys
+		    cp -p ${SRCDIR}/sys/sys/queue.h ${BRTOOLDIR}/compat/include/sys
+		    echo "CPPFLAGS+=-I${BRTOOLDIR}/compat/include" >> "${MKCONF}"
+		fi
 	fi
 
 	printoneconfig 'Cmd' "SRCDIR" "${SRCDIR}"
@@ -627,30 +629,31 @@ EOF
 
 	printoneconfig 'Mode' "${TARBALLMODE}" 'yes'
 
-	rm -f ${BRTOOLDIR}/toolchain-conf.mk
-	exec 3>&1 1>${BRTOOLDIR}/toolchain-conf.mk
-	printf 'BUILDRUMP_TOOL_CFLAGS=%s\n' "${EXTRA_CFLAGS}"
-	printf 'BUILDRUMP_TOOL_CXXFLAGS=%s\n' "${EXTRA_CFLAGS}"
-	printf 'BUILDRUMP_TOOL_CPPFLAGS=-D__NetBSD__ %s %s\n' \
-	    "${EXTRA_CPPFLAGS}" "${RUMPKERN_UNDEF}"
-	exec 1>&3 3>&-
+	if [ ${RUMPKERNEL} = "netbsd" ] ; then
+		rm -f ${BRTOOLDIR}/toolchain-conf.mk
+		exec 3>&1 1>${BRTOOLDIR}/toolchain-conf.mk
+		printf 'BUILDRUMP_TOOL_CFLAGS=%s\n' "${EXTRA_CFLAGS}"
+		printf 'BUILDRUMP_TOOL_CXXFLAGS=%s\n' "${EXTRA_CFLAGS}"
+		printf 'BUILDRUMP_TOOL_CPPFLAGS=-D__NetBSD__ %s %s\n' \
+		       "${EXTRA_CPPFLAGS}" "${RUMPKERN_UNDEF}"
+		exec 1>&3 3>&-
 
-	chkcrt begins
-	chkcrt ends
-	chkcrt i
-	chkcrt n
+		chkcrt begins
+		chkcrt ends
+		chkcrt i
+		chkcrt n
 
-	# add vars from env last (so that they can be used for overriding)
-	cat >> "${MKCONF}" << EOF
+		# add vars from env last (so that they can be used for overriding)
+		cat >> "${MKCONF}" << EOF
 CPPFLAGS+=\${BUILDRUMP_CPPFLAGS}
 CFLAGS+=\${BUILDRUMP_CFLAGS}
 AFLAGS+=\${BUILDRUMP_AFLAGS}
 LDFLAGS+=\${BUILDRUMP_LDFLAGS}
 EOF
 
-	if ! ${KERNONLY}; then
-		echo >> "${MKCONF}"
-		cat >> "${MKCONF}" << EOF
+		if ! ${KERNONLY}; then
+		    echo >> "${MKCONF}"
+		    cat >> "${MKCONF}" << EOF
 # Support for NetBSD Makefiles which use <bsd.prog.mk>
 # It's mostly a question of erasing dependencies that we don't
 # expect to see
@@ -664,50 +667,158 @@ LIBC=
 LDFLAGS+= -L\${BUILDRUMP_STAGE}/usr/lib -Wl,-R${DESTDIR}/lib
 LDADD+= ${EXTRA_RUMPCOMMON} ${EXTRA_RUMPUSER} ${EXTRA_RUMPCLIENT}
 EOF
-		appendmkconf 'Probe' "${LD_AS_NEEDED}" LDFLAGS +
-		echo '.endif # PROG' >> "${MKCONF}"
+		    appendmkconf 'Probe' "${LD_AS_NEEDED}" LDFLAGS +
+		    echo '.endif # PROG' >> "${MKCONF}"
+		fi
+
+		# skip the zlib tests run by "make tools", since we don't need zlib
+		# and it's only required by one tools autoconf script.  Of course,
+		# the fun bit is that autoconf wants to use -lz internally,
+		# so we provide some foo which macquerades as libz.a.
+		export ac_cv_header_zlib_h=yes
+		echo 'int gzdopen(int); int gzdopen(int v) { return 0; }' > fakezlib.c
+		${HOST_CC} -o libz.a -c fakezlib.c
+		rm -f fakezlib.c
+
+		# Run build.sh.  Use some defaults.
+		# The html pages would be nice, but result in too many broken
+		# links, since they assume the whole NetBSD man page set to be present.
+		cd ${SRCDIR}
+
+		# create user-usable wrapper script
+		makemake ${BRTOOLDIR}/rumpmake ${BRTOOLDIR}/dest makewrapper
+
+		# create wrapper script to be used during buildrump.sh, plus tools
+		makemake ${RUMPMAKE} ${OBJDIR}/dest.stage tools
+
+		# Just set no MSI in imacros universally now.
+		# Need to:
+		#   a) migrate more defines there
+		#   b) set no MSI only when necessary
+		printf '#define NO_PCI_MSI_MSIX\n' > ${BRIMACROS}.building
+
+		unset ac_cv_header_zlib_h
+
+		# tool build done.  flip mk.conf name so that it gets picked up
+		omkconf="${MKCONF}"
+		MKCONF="${mkconf_final}"
+		mv "${omkconf}" "${MKCONF}"
+		unset omkconf mkconf_final
+
+		# set new BRIMACROS only if the contents change (avoids
+		# full rebuild, since every file in the rump kernel depends on the
+		# contents of BRIMACROS
+		if ! diff "${BRIMACROS}" "${BRIMACROS}.building" > /dev/null 2>&1; then
+		    mv "${BRIMACROS}.building" "${BRIMACROS}"
+		fi
+	fi
+}
+
+makelinuxtools ()
+{
+
+	checkcheckout
+
+	probeld
+	probenm
+	probear
+	${HAVECXX} && probecxx
+
+	cd ${OBJDIR}
+
+	# Create mk.conf.  Create it under a temp name first so as to
+	# not affect the tool build with its contents
+	MKCONF="${BRTOOLDIR}/mk.conf.building"
+	> "${MKCONF}"
+	mkconf_final="${BRTOOLDIR}/mk.conf"
+	> ${mkconf_final}
+
+	${KERNONLY} || probe_rumpuserbits
+
+	checkcompiler
+
+	#
+	# Create external toolchain wrappers.
+	mkdir -p ${BRTOOLDIR}/bin || die "cannot create ${BRTOOLDIR}/bin"
+	for x in CC AR NM OBJCOPY; do
+		maketoolwrapper true $x
+	done
+	for x in AS CXX LD OBJDUMP RANLIB READELF SIZE STRINGS STRIP; do
+		maketoolwrapper false $x
+	done
+
+	# create a cpp wrapper, but run it via cc -E
+	if [ "${CC_FLAVOR}" = 'clang' ]; then
+		cppname=clang-cpp
+	else
+		cppname=cpp
+	fi
+	tname=${BRTOOLDIR}/bin/${MACHINE_GNU_ARCH}--${RUMPKERNEL}${TOOLABI}-${cppname}
+	printf '#!/bin/sh\n\nexec %s -E -x c "${@}"\n' ${CC} > ${tname}
+	chmod 755 ${tname}
+
+	for x in 1 2 3; do
+		! ${HOST_CC} -o ${BRTOOLDIR}/bin/brprintmetainfo \
+		    -DSTATHACK${x} ${BRDIR}/brlib/utils/printmetainfo.c \
+		    >/dev/null 2>&1 || break
+	done
+	[ -x ${BRTOOLDIR}/bin/brprintmetainfo ] \
+	    || die failed to build brprintmetainfo
+
+	${HOST_CC} -o ${BRTOOLDIR}/bin/brrealpath \
+	    ${BRDIR}/brlib/utils/realpath.c || die failed to build brrealpath
+
+	printoneconfig 'Cmd' "SRCDIR" "${SRCDIR}"
+	printoneconfig 'Cmd' "DESTDIR" "${DESTDIR}"
+	printoneconfig 'Cmd' "OBJDIR" "${OBJDIR}"
+	printoneconfig 'Cmd' "BRTOOLDIR" "${BRTOOLDIR}"
+
+	appendmkconf 'Cmd' "${RUMP_DIAGNOSTIC:-}" "RUMP_DIAGNOSTIC"
+	appendmkconf 'Cmd' "${RUMP_DEBUG:-}" "RUMP_DEBUG"
+	appendmkconf 'Cmd' "${RUMP_LOCKDEBUG:-}" "RUMP_LOCKDEBUG"
+	appendmkconf 'Cmd' "${DBG:-}" "DBG"
+	printoneconfig 'Cmd' "make -j[num]" "-j ${JNUM}"
+
+	if ${KERNONLY}; then
+		appendmkconf Cmd yes RUMPKERN_ONLY
 	fi
 
-	# skip the zlib tests run by "make tools", since we don't need zlib
-	# and it's only required by one tools autoconf script.  Of course,
-	# the fun bit is that autoconf wants to use -lz internally,
-	# so we provide some foo which macquerades as libz.a.
-	export ac_cv_header_zlib_h=yes
-	echo 'int gzdopen(int); int gzdopen(int v) { return 0; }' > fakezlib.c
-	${HOST_CC} -o libz.a -c fakezlib.c
-	rm -f fakezlib.c
-
-	# Run build.sh.  Use some defaults.
-	# The html pages would be nice, but result in too many broken
-	# links, since they assume the whole NetBSD man page set to be present.
-	cd ${SRCDIR}
-
-	# create user-usable wrapper script
-	makemake ${BRTOOLDIR}/rumpmake ${BRTOOLDIR}/dest makewrapper
-
-	# create wrapper script to be used during buildrump.sh, plus tools
-	makemake ${RUMPMAKE} ${OBJDIR}/dest.stage tools
-
-	# Just set no MSI in imacros universally now.
-	# Need to:
-	#   a) migrate more defines there
-	#   b) set no MSI only when necessary
-	printf '#define NO_PCI_MSI_MSIX\n' > ${BRIMACROS}.building
-
-	unset ac_cv_header_zlib_h
-
-	# tool build done.  flip mk.conf name so that it gets picked up
-	omkconf="${MKCONF}"
-	MKCONF="${mkconf_final}"
-	mv "${omkconf}" "${MKCONF}"
-	unset omkconf mkconf_final
-
-	# set new BRIMACROS only if the contents change (avoids
-	# full rebuild, since every file in the rump kernel depends on the
-	# contents of BRIMACROS
-	if ! diff "${BRIMACROS}" "${BRIMACROS}.building" > /dev/null 2>&1; then
-		mv "${BRIMACROS}.building" "${BRIMACROS}"
+	if ${KERNONLY} && ! cppdefines __NetBSD__; then
+		appendmkconf 'Cmd' '-D__NetBSD__' 'CPPFLAGS' +
+		appendmkconf 'Probe' "${RUMPKERN_UNDEF}" 'CPPFLAGS' +
+	else
+		appendmkconf 'Probe' "${RUMPKERN_UNDEF}" "RUMPKERN_UNDEF"
 	fi
+	appendmkconf 'Probe' "${RUMP_CURLWP:-}" 'RUMP_CURLWP' ?
+	appendmkconf 'Probe' "${CTASSERT:-}" "CPPFLAGS" +
+	appendmkconf 'Probe' "${RUMP_VIRTIF:-}" "RUMP_VIRTIF"
+	appendmkconf 'Probe' "${EXTRA_CWARNFLAGS}" "CWARNFLAGS" +
+	appendmkconf 'Probe' "${EXTRA_LDFLAGS}" "LDFLAGS" +
+	appendmkconf 'Probe' "${EXTRA_CPPFLAGS}" "CPPFLAGS" +
+	appendmkconf 'Probe' "${EXTRA_CFLAGS}" "BUILDRUMP_CFLAGS"
+	appendmkconf 'Probe' "${EXTRA_AFLAGS}" "BUILDRUMP_AFLAGS"
+	_tmpvar=
+	for x in ${EXTRA_RUMPUSER} ${EXTRA_RUMPCOMMON}; do
+		appendvar _tmpvar "${x#-l}"
+	done
+	appendmkconf 'Probe' "${_tmpvar}" "RUMPUSER_EXTERNAL_DPLIBS" +
+	_tmpvar=
+	for x in ${EXTRA_RUMPCLIENT} ${EXTRA_RUMPCOMMON}; do
+		appendvar _tmpvar "${x#-l}"
+	done
+	appendmkconf 'Probe' "${_tmpvar}" "RUMPCLIENT_EXTERNAL_DPLIBS" +
+	appendmkconf 'Probe' "${LDSCRIPT:-}" "RUMP_LDSCRIPT"
+	appendmkconf 'Probe' "${SHLIB_MKMAP:-}" 'SHLIB_MKMAP'
+	appendmkconf 'Probe' "${SHLIB_WARNTEXTREL:-}" "SHLIB_WARNTEXTREL"
+	appendmkconf 'Probe' "${MKSTATICLIB:-}"  "MKSTATICLIB"
+	appendmkconf 'Probe' "${MKPIC:-}"  "MKPIC"
+	appendmkconf 'Probe' "${MKSOFTFLOAT:-}"  "MKSOFTFLOAT"
+	appendmkconf 'Probe' $(${HAVECXX} && echo yes || echo no) _BUILDRUMP_CXX
+
+	printoneconfig 'Mode' "${TARBALLMODE}" 'yes'
+
+	CC=${BRTOOLDIR}/bin/${MACHINE_GNU_ARCH}--${RUMPKERNEL}${TOOLABI}-gcc
+
 }
 
 makemake ()
@@ -829,11 +940,26 @@ makelinuxbuild ()
 		VERBOSE="V=1"
 	fi
 
+	CROSS=$(${CC} -dumpmachine)
+	if [ ${CROSS} = "$(gcc -dumpmachine)" ]
+	then
+		CROSS=
+	else
+		CROSS=${CROSS}-
+	fi
+
 	set -e
+	set -x
+	mkdir -p ${OBJDIR}/lkl-linux
 	cd tools/lkl
-	rm -f lib/lkl.o lib/liblinux.a lib/liblkl-hijack.so
-	make RUMP_PREFIX=${OBJDIR}/dest.stage/ -j ${JNUM} ${VERBOSE}
+	rm -f ${OBJDIR}/lkl-linux/tools/lkl/lib/lkl.o
+	export RUMP_PREFIX=${RUMPSRC}/sys/rump # ${OBJDIR}/dest.stage/
+	export RUMP_INCLUDE=${RUMPSRC}/sys/rump/include #${OBJDIR}/dest.stage/usr/include
+	make CROSS_COMPILE=${CROSS} rumprun=yes -j ${JNUM} ${VERBOSE} O=${OBJDIR}/lkl-linux/ # FIXME: not supported yet O=${OBJDIR}/lkl-linux/
+	cd ../../
+	make CROSS_COMPILE=${CROSS} headers_install ARCH=lkl O=${RROBJ}/rumptools/dest
 	set +e
+	set +x
 }
 
 linuxtest ()
@@ -845,13 +971,15 @@ linuxtest ()
 makeinstall ()
 {
 
-	# ensure we run this in a directory that does not have a
-	# Makefile that could confuse rumpmake
-	stage=$(cd ${BRTOOLDIR} && ${RUMPMAKE} -V '${BUILDRUMP_STAGE}')
-	(cd ${stage}/usr ; tar -cf - .) | (cd ${DESTDIR} ; tar -xf -)
-
-	if ${BUILDLINUX}; then
-		make install DESTDIR=${DESTDIR} -C ${LINUX_SRCDIR}/tools/lkl/
+	if [ ${RUMPKERNEL} = "netbsd" ] ; then
+		# ensure we run this in a directory that does not have a
+		# Makefile that could confuse rumpmake
+		stage=$(cd ${BRTOOLDIR} && ${RUMPMAKE} -V '${BUILDRUMP_STAGE}')
+		(cd ${stage}/usr ; tar -cf - .) | (cd ${DESTDIR} ; tar -xf -)
+	else
+		export RUMP_PREFIX=${RUMPSRC}/sys/rump # ${OBJDIR}/dest.stage/
+		export RUMP_INCLUDE=${RUMPSRC}/sys/rump/include #${OBJDIR}/dest.stage/usr/include
+		make rumprun=yes install DESTDIR=${DESTDIR} -C ${LINUX_SRCDIR}/tools/lkl/ O=${OBJDIR}/lkl-linux/
 	fi
 
 }
@@ -1366,7 +1494,7 @@ parseargs ()
 	NOISE=2
 	debugginess=0
 	KERNONLY=false
-	BUILDLINUX=false
+	RUMPKERNEL=netbsd
 	OBJDIR=./obj
 	DESTDIR=./rump
 	SRCDIR=./src
@@ -1435,7 +1563,7 @@ parseargs ()
 			KERNONLY=true
 			;;
 		l)
-			BUILDLINUX=true
+			RUMPKERNEL=linux
 			if [ ! -z ${OPTARG} ]; then
 				LINUX_SRCDIR=${OPTARG}
 			fi
@@ -1483,7 +1611,7 @@ parseargs ()
 	allcmds='checkout checkoutcvs checkoutgit probe tools build install
 	    tests fullbuild kernelheaders linuxbuild'
 	fullbuildcmds="tools build install"
-	if ${BUILDLINUX} ; then
+	if [ ${RUMPKERNEL} = "linux" ] ; then
 	    fullbuildcmds="${fullbuildcmds} linuxbuild"
 	fi
 
@@ -1532,7 +1660,7 @@ parseargs ()
 		docheckout=true
 		checkoutstyle=cvs
 	fi
-	if ${docheckout} && ${BUILDLINUX} ; then
+	if ${docheckout} && ${RUMPKERNEL} = "linux" ; then
 		docheckout=true
 		checkoutstyle=linux-git
 	fi
@@ -1666,7 +1794,7 @@ if ${doprobe} || ${dotools} || ${dobuild} || ${dokernelheaders} \
 
 	${KERNONLY} || evalplatform
 
-	export BUILDLINUX
+	export RUMPKERNEL
 	export LINUX_SRCDIR
 	${doprobe} && writeproberes
 	${dotools} && maketools
@@ -1683,7 +1811,7 @@ if ${doprobe} || ${dotools} || ${dobuild} || ${dokernelheaders} \
 			alltests
 		fi
 
-		if ${BUILDLINUX}; then
+		if [ ${RUMPKERNEL} = "linux" ] ; then
 			linuxtest
 		fi
 	fi
